requirements.txt
# Requirements for the project
pydantic
streamlit
cohere
scikit-learn
matplotlib
plotly
python-docx
pdfplumber
reportlab
